reflesh directory
