+++ .gitignore
@@ -1,2 +1,3 @@
 /target
+rtm-mpi/target/
 .DS_Store
