+++ rtm-lib/Cargo.toml
@@ -1,22 +1,16 @@
 [package]
-name = "rtm"
+name = "rtm-lib"
 version = "1.1.3"
-authors = ["th 2023"]
 edition = "2021"
-license = "MIT"
-repository = "https://github.com/th2ch-g/rust-top_message"
+authors = ["th 2023"]
 description = "rust-top_message: A tool which display message on top"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rtm-lib = { path = "./rtm-lib/" }
+clap = { version = "3.2.22", features = ["derive"]  }
+chrono = "0.4.22"
+rand = "0.8.5"
 
 [dev-dependencies]
 
-[profile.release]
-lto = true
-strip = true
-
-[profile.dev]
-opt-level = 3
