+++ rtm-lib/src/check.rs
