+++ rtm-lib/src/common.rs
