+++ rtm-lib/src/multiple2.rs
