+++ rtm-lib/src/vertical.rs
