+++ rtm-mpi/Cargo.toml
@@ -1,22 +1,19 @@
 [package]
-name = "rtm"
+name = "rtm-mpi"
 version = "1.1.3"
-authors = ["th 2023"]
 edition = "2021"
-license = "MIT"
-repository = "https://github.com/th2ch-g/rust-top_message"
-description = "rust-top_message: A tool which display message on top"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rtm-lib = { path = "./rtm-lib/" }
+rtm-lib = { path = "../rtm-lib/" }
+mpi = "0.6.0"
 
 [dev-dependencies]
 
 [profile.release]
 lto = true
-strip = true
 
 [profile.dev]
 opt-level = 3
+
